@@ -66,18 +66,21 @@
     "@lingui/react": "3.17.2",
     "@mui/lab": "5.0.0-alpha.122",
     "@mui/material": "5.11.12",
-    "@next/env": "13.2.4",
+    "@next/env": "13.3.0",
+    "client-only": "^0.0.1",
     "concurrently": "7.6.0",
     "cross-env": "^7.0.3",
     "dotenv": "16.0.3",
+    "encoding": "^0.1.13",
     "framer-motion": "10.2.4",
     "graphql": "16.6.0",
-    "next": "13.2.4",
+    "next": "13.3.0",
     "next-pwa": "5.6.0",
     "next-sitemap": "4.0.5",
     "patch-package": "6.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "server-only": "^0.0.1",
     "sharp": "0.31.3",
     "webpack": "5.76.0"
   },
@@ -90,8 +93,8 @@
     "@types/react-dom": "18.0.11",
     "eslint": "8.35.0",
     "prettier": "2.8.4",
-    "type-fest": "^3.4.0",
-    "typescript": "4.9.5"
+    "typescript": "5.0.4",
+    "typescript": "5.0.4"
   },
   "browserslist": [
     "> 1% in alt-EU",
