@@ -16,13 +16,13 @@
     "@graphcommerce/next-ui": "6.1.1-canary.2",
     "@mui/material": "5.11.12",
     "framer-motion": "10.2.4",
-    "next": "13.2.4",
+    "next": "13.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "license": "MIT",
   "devDependencies": {
-    "typescript": "4.9.5"
+    "typescript": "5.0.4"
   },
   "prettier": "@graphcommerce/prettier-config-pwa",
   "eslintConfig": {
