@@ -25,7 +25,7 @@
     "@graphcommerce/eslint-config-pwa": "6.1.1-canary.2",
     "@graphcommerce/prettier-config-pwa": "6.1.1-canary.2",
     "@graphcommerce/typescript-config-pwa": "6.1.1-canary.2",
-    "typescript": "4.9.5"
+    "typescript": "5.0.4"
   },
   "sideEffects": false,
   "prettier": "@graphcommerce/prettier-config-pwa",
