@@ -14,12 +14,12 @@
     "@graphcommerce/next-config": "6.1.1-canary.2",
     "@graphcommerce/next-ui": "6.1.1-canary.2",
     "framer-motion": "^10.2.4",
-    "next": "13.2.4",
+    "next": "13.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "license": "MIT",
   "devDependencies": {
-    "typescript": "4.9.5"
+    "typescript": "5.0.4"
   }
 }
