@@ -26,7 +26,7 @@
     "@types/relay-compiler": "8.0.3",
     "jest": "29.5.0",
     "ts-jest": "29.0.5",
-    "typescript": "4.9.5"
+    "typescript": "5.0.4"
   },
   "prettier": "@graphcommerce/prettier-config-pwa",
   "eslint": {
