@@ -1,9 +1,9 @@
-export * from './Backstory'
-export * from './Feature'
-export * from './FeatureBoxed'
-export * from './Grid'
-export * from './Related'
-export * from './Reviews'
-export * from './Specs'
-export * from './Swipeable'
-export * from './Upsells'
+export { Backstory } from './Backstory'
+export { Feature } from './Feature'
+export { FeatureBoxed } from './FeatureBoxed'
+export { Grid } from './Grid'
+export { Related } from './Related'
+export { Reviews } from './Reviews'
+export { Specs } from './Specs'
+export { Swipeable } from './Swipeable'
+export { Upsells } from './Upsells'
