@@ -1,4 +1,3 @@
-export * from './RowProduct/RowProduct'
-
-// The actual renderer
-export * from './RowRenderer'
+export { RowProduct } from './RowProduct/RowProduct'
+export { RowRenderer } from './RowRenderer'
+export type { PageProps } from './RowRenderer'
