@@ -3,7 +3,7 @@ import { AnimatePresence, m } from 'framer-motion'
 import { requestIdleCallback, cancelIdleCallback } from 'next/dist/client/request-idle-callback'
 import { PrivateRouteInfo } from 'next/dist/shared/lib/router/router'
 import { AppPropsType } from 'next/dist/shared/lib/utils'
-import { NextRouter, Router } from 'next/router'
+import { NextRouter, Router } from 'next/compat/router'
 import { useEffect, useRef, useState } from 'react'
 import { pageContext } from '../context/pageContext'
 import type { PageComponent, PageItem, UpPage } from '../types'
