@@ -1,5 +1,5 @@
 import { createContext } from 'react'
 import { PageContext } from '../types'
 
-export const pageRouterContext = createContext(undefined as unknown as PageContext)
+export const pageRouterContext = createContext<PageContext | undefined>(undefined)
 pageRouterContext.displayName = 'PageRouterContext'
