@@ -10,7 +10,7 @@ import {
 import Link, { Container } from '@mui/material'
 import { m } from 'framer-motion'
 import { GetStaticPathsResult, GetStaticProps } from 'next'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 import React, { useState } from 'react'
 import { Grid } from '../../components/Grid'
 import { StackDebug } from '../../components/StackedDebugger'
