@@ -3,5 +3,5 @@ import { pageRouterContext } from '../context/pageRouterContext'
 
 /** Get the upUrl of the previous page */
 export function usePrevUp() {
-  return useContext(pageRouterContext).prevUp
+  return useContext(pageRouterContext)?.prevUp
 }
