@@ -3,7 +3,7 @@ import { useApolloClient, useQuery } from '@graphcommerce/graphql'
 import { productLink, ProductLinkProps } from '@graphcommerce/magento-product'
 import { StoreConfigDocument } from '@graphcommerce/magento-store'
 import { Button, ButtonProps, SxProps, Theme } from '@mui/material'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 import React from 'react'
 import { RouteDocument } from '../../Route.gql'
 import { ButtonLinkProps } from './getButtonLinkProps'
