@@ -6,7 +6,7 @@ import {
 } from '@graphcommerce/magento-cart-payment-method'
 import { FormRow, InputCheckmark } from '@graphcommerce/next-ui'
 import { TextField } from '@mui/material'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 import { useAdyenCartLock } from '../../hooks/useAdyenCartLock'
 import { useAdyenPaymentMethod } from '../../hooks/useAdyenPaymentMethod'
 import {
@@ -67,7 +67,10 @@ export function HppOptions(props: PaymentOptionsProps) {
 
   if (!conf?.issuers?.length) return <form onSubmit={submit} noValidate />
 
-  /** This is the form that the user can fill in. In this case we don't wat the user to fill in anything. */
+  /**
+   * This is the form that the user can fill in. In this case we don't wat the user to fill in
+   * anything.
+   */
   return (
     <form key={key} onSubmit={submit} noValidate>
       {conf?.issuers && (
