@@ -3,19 +3,11 @@ export * from './JsonLdProduct/jsonLdProduct'
 export * from './ProductAddToCart'
 export * from './ProductCustomizable'
 export * from './ProductFiltersPro'
-export * from './ProductList/ProductList.gql'
 export * from './ProductListCount/ProductListCount'
 export * from './ProductListFilters'
-export * from './ProductListFiltersContainer/ProductListFiltersContainer'
+export * from './ProductListFiltersContainer'
 export * from './ProductListItem/ProductListItem'
-export * from './ProductListItems/filteredProductList'
-export * from './ProductListItems/filterTypes'
-export * from './ProductListItems/getFilterTypes'
-export * from './ProductListItems/ProductListItems'
-export * from './ProductListItems/ProductListItems.gql'
-export * from './ProductListItems/ProductListItemsBase'
-export * from './ProductListItems/ProductListParamsProvider'
-export * from './ProductListItems/renderer'
+export * from './ProductListItems'
 export * from './ProductListLink/ProductListLink'
 export * from './ProductListPagination/ProductListPagination'
 export * from './ProductListSort'
@@ -25,12 +17,9 @@ export * from './ProductPageDescription/ProductPageDescription'
 export * from './ProductPageGallery/ProductPageGallery'
 export * from './ProductPageMeta/ProductPageMeta'
 export * from './ProductPagePrice'
-export * from './ProductRelated/RelatedProducts.gql'
 export * from './ProductShortDescription/ProductShortDescription'
-export * from './ProductShortDescription/ProductShortDescription.gql'
 export * from './ProductSidebarDelivery/ProductSidebarDelivery'
 export * from './ProductSpecs/ProductSpecs'
 export * from './ProductStaticPaths/getProductStaticPaths'
 export * from './ProductStaticPaths/getSitemapPaths'
-export * from './ProductUpsells/UpsellProducts.gql'
 export * from './ProductWeight/ProductWeight'
