@@ -1,5 +1,5 @@
 import { ListItemButton, ListItemButtonProps, ListItemText } from '@mui/material'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 import { NextLink } from '../Theme'
 
 export type MenuFabItemProps = Omit<ListItemButtonProps<'a'>, 'href' | 'button'> & {
@@ -9,7 +9,7 @@ export type MenuFabItemProps = Omit<ListItemButtonProps<'a'>, 'href' | 'button'>
 export function MenuFabItem(props: MenuFabItemProps) {
   const { href, children, sx = [], ...listItemProps } = props
   const hrefString = href.toString()
-  const path = useRouter().asPath.split('?')[0]
+  const path = useRouter()?.asPath.split('?')[0]
   const active = hrefString === '/' ? path === hrefString : path.startsWith(hrefString)
 
   return (
