@@ -1,7 +1,7 @@
 import { useMotionValueValue } from '@graphcommerce/framer-utils'
 import { Fab, styled, Box, SxProps, Theme, FabProps } from '@mui/material'
 import { m } from 'framer-motion'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 import React, { useEffect } from 'react'
 import { IconSvg } from '../../IconSvg'
 import { useScrollY } from '../../Layout/hooks/useScrollY'
@@ -37,9 +37,9 @@ export function NavigationFab(props: NavigationFabProps) {
 
   useEffect(() => {
     const clear = () => setOpenEl(null)
-    router.events.on('routeChangeStart', clear)
-    return () => router.events.off('routeChangeStart', clear)
-  }, [router.events])
+    router?.events.on('routeChangeStart', clear)
+    return () => router?.events.off('routeChangeStart', clear)
+  }, [router?.events])
 
   const fabIconSize = useFabSize('responsive')
 
