@@ -1,4 +1,4 @@
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/compat/router'
 
 export const storefrontAll = import.meta.graphCommerce.storefront
 
@@ -11,7 +11,7 @@ export const storefrontConfigDefault = () =>
 
 /** Automatically selects the correct storefront config based on the current locale */
 export function useStorefrontConfig(locale?: string | undefined) {
-  const routerLocale = useRouter().locale
+  const routerLocale = useRouter()?.locale
   const config = storefrontConfig(locale ?? routerLocale)
   if (!config) throw Error(`No storefront config found for locale ${locale}`)
   return config
