@@ -1,5 +1,5 @@
-import { loadConfig } from '../loadConfig'
-import { exportConfigToEnv } from '../utils/exportConfigToEnv'
+import { loadConfig } from '../loadConfig.js'
+import { exportConfigToEnv } from '../utils/exportConfigToEnv.js'
 
 // eslint-disable-next-line @typescript-eslint/require-await
 export async function exportConfig() {
